separated non-Jena data related stuff out of TriplePatternFragmentImpl into TriplePatternFragmentBase
